Microservicios/Microservicios/autenticacion/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
marshmallow==3.21.1
Flask-Marshmallow==0.15.0
Flask-JWT-Extended==4.6.0
sendgrid==6.11.0
redis==5.0.3
requests==2.31.0
